# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/edge/importtemplate": {
            "post": {
                "description": "匯入範本",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "範本列表"
                ],
                "parameters": [
                    {
                        "description": "範本關聯表資料",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/templaterouter.editTmpRel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/templaterouter.errMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/templaterouter.errMsg"
                        }
                    }
                }
            }
        },
...
        "/edge/model": {
            "get": {
                "description": "取得設備模型資料",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "範本列表"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/templaterouter.errMsg"
                        }
                    }
                }
            }
        },
...
        "/edge/template": {
            "get": {
                "description": "取得範本列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "範本列表"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/nodeattr.Template"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/templaterouter.errMsg"
                        }
                    }
                }
            },
            "put": {
                "description": "編輯範本列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "範本列表"
                ],
                "parameters": [
                    {
                        "description": "編輯資料",
                        "name": "editinfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/templaterouter.updateTemp"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/templaterouter.errMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/templaterouter.errMsg"
                        }
                    }
                }
            },
            "post": {
                "description": "新增範本列表 \u003cbr\u003e使用複製範本功能，id填入來源範本id，name填入新範本名稱",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "範本列表"
                ],
                "parameters": [
                    {
                        "description": "範本資料",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/templaterouter.tmpInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/templaterouter.errMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/templaterouter.errMsg"
                        }
                    }
                }
            },
            "delete": {
                "description": "刪除範本列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "範本列表"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "範本名稱",
                        "name": "id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/templaterouter.errMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/templaterouter.errMsg"
                        }
                    }
                }
            }
        },
        "/edge/temprel": {
            "get": {
                "description": "取得範本關聯資訊",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "範本列表"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "範本ID",
                        "name": "id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/nodeattr.TemplateRel"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/templaterouter.errMsg"
                        }
                    }
                }
            },
            "put": {
                "description": "編輯範本關聯資訊",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "範本列表"
                ],
                "parameters": [
                    {
                        "description": "範本關聯表資料",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/templaterouter.editTmpRel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/templaterouter.errMsg"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/templaterouter.errMsg"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "nodeattr.Template": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "modifyTime": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "nodeattr.TemplateRel": {
            "type": "object",
            "properties": {
                "convertFunc": {
                    "type": "string"
                },
                "dstBrowseName": {
                    "type": "string"
                },
                "dstNodeId": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "referenceBrowseName1": {
                    "type": "string"
                },
                "referenceBrowseName2": {
                    "type": "string"
                },
                "referenceBrowseName3": {
                    "type": "string"
                },
                "srcBrowseName": {
                    "type": "string"
                },
                "srcNamespace": {
                    "type": "integer"
                },
                "srcNodeId": {
                    "type": "string"
                },
                "srcUnit": {
                    "type": "string"
                },
                "templateId": {
                    "type": "integer"
                }
            }
        },
        "templaterouter.editTmpRel": {
            "type": "object",
            "properties": {
                "editContent": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/nodeattr.TemplateRel"
                    }
                },
                "name": {
                    "type": "string"
                },
                "templateId": {
                    "type": "integer"
                }
            }
        },
        "templaterouter.errMsg": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
        "templaterouter.tmpInfo": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "modifyTime": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "templaterouter.updateTemp": {
            "type": "object",
            "properties": {
                "editContent": {
                    "$ref": "#/definitions/nodeattr.Template"
                },
                "id": {
                    "type": "integer"
                }
            }
        }
    }
}